Fixed modernity issues with 1.9 and dependencies
Updated clojure dep to 1.9, and fixed a subsequent
error within the reactive macro definitions that
spec found.  Turns out, the name-symbols for the
anonymous functions weren't actually quoted...
so the actual ns-qualified var was being used
and causing a problem!  Fix was to quote said
symbols inside the macros.

Also updated data.avl dep to latest public
release 0.0.17 just because.